requirements.txt
pandas
matplotlib
openpyxl
schedule
psycopg2-binary
slack_sdk
python-dotenv